Merge Release 2.4.0

+++ sqswatcher/sqswatcher.py
@@ -15,7 +15,6 @@
 from future.moves.collections import OrderedDict
 
 import collections
-import ConfigParser
 import itertools
 import json
 import logging
@@ -24,9 +23,21 @@ import time
 import boto3
 from botocore.config import Config
 from botocore.exceptions import ClientError
+from configparser import ConfigParser
 from retrying import retry
 
-from common.utils import CriticalError, get_asg_name, get_asg_settings, load_module
+from common.time_utils import seconds
+from common.utils import (
+    CriticalError,
+    get_asg_name,
+    get_asg_settings,
+    get_compute_instance_type,
+    get_instance_properties,
+    load_module,
+)
+
+LOOP_TIME = 30
+CLUSTER_PROPERTIES_REFRESH_INTERVAL = 180
 
 
 class QueryConfigError(Exception):
@@ -37,8 +48,7 @@ log = logging.getLogger(__name__)
 
 
 SQSWatcherConfig = collections.namedtuple(
-    "SQSWatcherConfig",
-    ["region", "scheduler", "sqsqueue", "table_name", "cluster_user", "proxy_config", "max_queue_size", "stack_name"],
+    "SQSWatcherConfig", ["region", "scheduler", "sqsqueue", "table_name", "cluster_user", "proxy_config", "stack_name"]
 )
 
 Host = collections.namedtuple("Host", ["instance_id", "hostname", "slots"])
@@ -55,7 +65,7 @@ def _get_config():
     config_file = "/etc/sqswatcher.cfg"
     log.info("Reading %s", config_file)
 
-    config = ConfigParser.RawConfigParser()
+    config = ConfigParser()
     config.read(config_file)
     if config.has_option("sqswatcher", "loglevel"):
         lvl = logging._levelNames[config.get("sqswatcher", "loglevel")]
@@ -66,7 +76,6 @@ def _get_config():
     sqsqueue = config.get("sqswatcher", "sqsqueue")
     table_name = config.get("sqswatcher", "table_name")
     cluster_user = config.get("sqswatcher", "cluster_user")
-    max_queue_size = int(config.get("sqswatcher", "max_queue_size"))
     stack_name = config.get("sqswatcher", "stack_name")
 
     _proxy = config.get("sqswatcher", "proxy")
@@ -76,19 +85,16 @@ def _get_config():
 
     log.info(
         "Configured parameters: region=%s scheduler=%s sqsqueue=%s table_name=%s cluster_user=%s "
-        "proxy=%s max_queue_size=%d stack_name=%s",
+        "proxy=%s stack_name=%s",
         region,
         scheduler,
         sqsqueue,
         table_name,
         cluster_user,
         _proxy,
-        max_queue_size,
         stack_name,
     )
-    return SQSWatcherConfig(
-        region, scheduler, sqsqueue, table_name, cluster_user, proxy_config, max_queue_size, stack_name
-    )
+    return SQSWatcherConfig(region, scheduler, sqsqueue, table_name, cluster_user, proxy_config, stack_name)
 
 
 @retry(stop_max_attempt_number=3, wait_fixed=5000)
@@ -116,7 +122,7 @@ def _get_sqs_queue(region, queue_name, proxy_config):
 @retry(
     stop_max_attempt_number=3,
     wait_fixed=5000,
-    retry_on_exception=lambda exception: not isinstance(exception, CriticalError)
+    retry_on_exception=lambda exception: not isinstance(exception, CriticalError),
 )
 def _get_ddb_table(region, table_name, proxy_config):
     """
@@ -166,7 +172,22 @@ def _requeue_message(queue, message):
     :param queue: the queue where to send the message
     :param message: the message to requeue
     """
-    queue.send_message(MessageBody=message.body, DelaySeconds=60)
+    max_retries = 2
+    message_body = json.loads(message.body)
+    if "TTL" not in message_body:
+        message_body["TTL"] = max_retries
+    else:
+        message_body["TTL"] = message_body["TTL"] - 1
+        if message_body["TTL"] < 1:
+            log.warning(
+                "Discarding the following message since exceeded the number of max retries (%d): %s",
+                max_retries,
+                message_body,
+            )
+            return
+    message_body_string = json.dumps(message_body)
+    log.warning("Re-queuing failed event %s", message_body_string)
+    queue.send_message(MessageBody=message_body_string, DelaySeconds=60)
 
 
 def _retrieve_all_sqs_messages(queue):
@@ -254,35 +275,44 @@ def _process_instance_terminate_event(message_attrs, message, table):
 
 
 def _process_sqs_messages(
-    update_events, scheduler_module, sqs_config, table, queue, max_cluster_size, update_max_cluster_size
+    update_events,
+    scheduler_module,
+    sqs_config,
+    table,
+    queue,
+    max_cluster_size,
+    instance_properties,
+    force_cluster_update,
 ):
     # Update the scheduler only when there are messages from the queue or
     # tha ASG max size got updated.
-    if not update_events and not update_max_cluster_size:
+    if not update_events and not force_cluster_update:
         return
 
     failed_events, succeeded_events = scheduler_module.update_cluster(
-        max_cluster_size, sqs_config.cluster_user, update_events
+        max_cluster_size, sqs_config.cluster_user, update_events, instance_properties
     )
 
-    for event in list(succeeded_events):
+    for event in update_events:
         try:
+            # Add an item to the table also in case of failures to handle host removal correctly
             if event.action == "ADD":
                 _retry_on_request_limit_exceeded(
                     lambda: table.put_item(Item={"instanceId": event.host.instance_id, "hostname": event.host.hostname})
                 )
-            elif event.action == "REMOVE":
+            # Remove item from table only in case of success
+            elif event.action == "REMOVE" and event in succeeded_events:
                 _retry_on_request_limit_exceeded(lambda: table.delete_item(Key={"instanceId": event.host.instance_id}))
             log.debug("Successfully processed event %s", event)
         except Exception as e:
             log.error(
                 "Failed when updating dynamo db table for instance %s with exception %s", event.host.instance_id, e
             )
-            failed_events.append(event)
-            succeeded_events.remove(event)
+            if event not in failed_events:
+                failed_events.append(event)
+                succeeded_events.remove(event)
 
     for event in failed_events:
-        log.warning("Re-queuing failed event %s", event)
         _requeue_message(queue, event.message)
 
     for event in itertools.chain(failed_events, succeeded_events):
@@ -292,7 +322,7 @@ def _process_sqs_messages(
 
 def _retrieve_max_cluster_size(sqs_config, asg_name, fallback):
     try:
-        _, _, max_size = get_asg_settings(sqs_config.region, sqs_config.proxy_config, asg_name, log)
+        _, _, max_size = get_asg_settings(sqs_config.region, sqs_config.proxy_config, asg_name)
         return max_size
     except Exception:
         return fallback
@@ -308,9 +338,30 @@ def _poll_queue(sqs_config, queue, table, asg_name):
     """
     scheduler_module = load_module("sqswatcher.plugins." + sqs_config.scheduler)
 
-    max_cluster_size = sqs_config.max_queue_size
+    max_cluster_size = None
+    instance_type = None
+    cluster_properties_refresh_timer = 0
     while True:
-        new_max_cluster_size = _retrieve_max_cluster_size(sqs_config, asg_name, max_cluster_size)
+        force_cluster_update = False
+        # dynamically retrieve max_cluster_size and compute_instance_type
+        if (
+            not max_cluster_size
+            or not instance_type
+            or cluster_properties_refresh_timer >= CLUSTER_PROPERTIES_REFRESH_INTERVAL
+        ):
+            cluster_properties_refresh_timer = 0
+            logging.info("Refreshing cluster properties")
+            new_max_cluster_size = _retrieve_max_cluster_size(sqs_config, asg_name, fallback=max_cluster_size)
+            new_instance_type = get_compute_instance_type(
+                sqs_config.region, sqs_config.proxy_config, sqs_config.stack_name, fallback=instance_type
+            )
+            force_cluster_update = new_max_cluster_size != max_cluster_size or new_instance_type != instance_type
+            if new_instance_type != instance_type:
+                instance_type = new_instance_type
+                instance_properties = get_instance_properties(sqs_config.region, sqs_config.proxy_config, instance_type)
+            max_cluster_size = new_max_cluster_size
+        cluster_properties_refresh_timer += LOOP_TIME
+
         messages = _retrieve_all_sqs_messages(queue)
         update_events = _parse_sqs_messages(messages, table)
         _process_sqs_messages(
@@ -319,14 +370,14 @@ def _poll_queue(sqs_config, queue, table, asg_name):
             sqs_config,
             table,
             queue,
-            new_max_cluster_size,
-            new_max_cluster_size != max_cluster_size,
+            max_cluster_size,
+            instance_properties,
+            force_cluster_update,
         )
-        max_cluster_size = new_max_cluster_size
-        time.sleep(30)
+        time.sleep(LOOP_TIME)
 
 
-@retry(wait_fixed=30000)
+@retry(wait_fixed=seconds(LOOP_TIME))
 def main():
     logging.basicConfig(level=logging.INFO, format="%(asctime)s %(levelname)s [%(module)s:%(funcName)s] %(message)s")
     log.info("sqswatcher startup")
@@ -335,7 +386,7 @@ def main():
         config = _get_config()
         queue = _get_sqs_queue(config.region, config.sqsqueue, config.proxy_config)
         table = _get_ddb_table(config.region, config.table_name, config.proxy_config)
-        asg_name = get_asg_name(config.stack_name, config.region, config.proxy_config, log)
+        asg_name = get_asg_name(config.stack_name, config.region, config.proxy_config)
 
         _poll_queue(config, queue, table, asg_name)
     except Exception as e:
